% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ft_data.R
\name{as.ft_data}
\alias{as.ft_data}
\title{Coerce directory of papers to ft_data object}
\usage{
as.ft_data(path = NULL)
}
\arguments{
\item{path}{cache path. if not given, we use the default
cache path. Default: \code{NULL}}
}
\value{
an object of class \code{ft_data}
}
\description{
create the same object that \code{\link[=ft_get]{ft_get()}} outputs
from your cached files - without having to run
\code{\link[=ft_get]{ft_get()}} again
}
\examples{
x <- as.ft_data()
# output lives underneath a special list index "cached" 
#   representing already present files
x$cached

y <- x \%>\% ft_collect()
y$cached$data$data
y \%>\% ft_chunks(what="title") \%>\% ft_tabularize()
x \%>\% ft_collect() \%>\% ft_text()
}
