% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltexttable.R
\docType{methods}
\name{as.fulltexttable}
\alias{as.fulltexttable}
\alias{as.fulltexttable,slice-method}
\title{Convert object to input for fulltext (table format).}
\usage{
as.fulltexttable(x, ...)

\S4method{as.fulltexttable}{slice}(x, display = c("none", "block"),
  headline = NULL, name = "")
}
\arguments{
\item{x}{The object to be converted.}

\item{...}{Placeholder for further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block"}

\item{headline}{A headline to prepend.}

\item{name}{An id inserted into tags.}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
library(polmineR)
library(fulltext)
use("polmineR")
p <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
ftab <- as.fulltexttable(p, headline = "Volker Kauder (CDU)", display = "block")
fulltext(ftab, box = FALSE)

sd <- crosstalk::SharedData$new(ftab)
fulltext::fulltext(sd)
}
