% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ftdmurl.R
\name{as_ftdmurl}
\alias{as_ftdmurl}
\alias{as_ftdmurl.ftdmurl}
\alias{as_ftdmurl.character}
\title{Coerce a url to a tdmurl with a specific type}
\usage{
as_ftdmurl(url, type, doi = NULL, member = NULL, intended_application = NULL)

\method{as_ftdmurl}{ftdmurl}(url, type, doi = NULL, member = NULL, intended_application = NULL)

\method{as_ftdmurl}{character}(url, type, doi = NULL, member = NULL, intended_application = NULL)
}
\arguments{
\item{url}{(character) A URL.}

\item{type}{(character) One of 'xml' (default), 'html', 'plain', 'pdf',
'unspecified', or 'all'}

\item{doi}{(character) A DOI, optional, default: \code{NULL}}

\item{member}{(character) Crossref member id. optional}

\item{intended_application}{(character) intended application string,
optional}
}
\description{
A tmd url is just a URL with some attributes to make it easier
to handle within other functions in this package.
}
\examples{
as_ftdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.xml",
   "xml")
as_ftdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.pdf",
   "pdf")
out <-
 as_ftdmurl("http://downloads.hindawi.com/journals/bmri/2014/201717.pdf",
   "pdf", "10.1155/2014/201717")
attributes(out)
identical(attr(out, "type"), "pdf")
}
