% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\alias{cache_options_set}
\alias{cache_options_get}
\title{Cache blobs of json, xml or pdfs of text from \code{ft_get()} function}
\usage{
cache_options_set(path = "fulltext", backend = "ext", overwrite = FALSE)

cache_options_get()
}
\arguments{
\item{path}{(character) End of directory path. Default: "fulltext".
See Details.}

\item{backend}{(character) Only "ext" supported for now.}

\item{overwrite}{(logical) overwrite cached file or not. Default: \code{FALSE}}
}
\description{
Cache blobs of json, xml or pdfs of text from \code{ft_get()} function
}
\section{Managing cached files}{

The dafault cache directory is \code{paste0(rappdirs::user_cache_dir(), "/R/fulltext")},
but you can set your own path using \code{cache_path_set()}

\code{cache_delete} only accepts 1 file name, while
\code{cache_delete_all} doesn't accept any names, but deletes all files.
For deleting many specific files, use \code{cache_delete} in a \code{\link[=lapply]{lapply()}}
type call
}

\section{Useful user functions for managing cached files}{

\itemize{
\item \code{ftxt_cache$list()} returns a character vector of full
path file names
\item \code{ftxt_cache$files()} returns file objects with metadata
\item \code{ftxt_cache$details()} returns files with details
\item \code{ftxt_cache$delete()} delete specific files
\item \code{ftxt_cache$delete_all()} delete all files, returns nothing
}
}

\examples{
\dontrun{
# Manage cached files with this object
ftxt_cache

# list files in cache
ftxt_cache$list()

# list details of files in cache
ftxt_cache$details()

# delete certain database files
# ftxt_cache$delete("file path")
# ftxt_cache$list()

# delete all files in cache
# ftxt_cache$delete_all()
# ftxt_cache$list()
}
}
