% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chunks.R
\name{chunks}
\alias{chunks}
\title{Extract chunks of data from articles.}
\usage{
chunks(x, what = "all")
}
\arguments{
\item{x}{Input article}

\item{what}{What to get, can be 1 to many. See Details.}
}
\value{
A list of output, one for each thing requested
}
\description{
Extract chunks of data from articles.
}
\details{
\itemize{
 \item title
 \item doi
 \item categories
 \item authors
 \item keywords
 \item body
 \item abstract
 \item executive_summary
 \item refs
 \item refs_dois
 \item publisher
 \item journal_meta
}
}
\examples{
\dontrun{
x <- ft_get(ids=c('10.7554/eLife.04251','10.7554/eLife.04986'), from='elife')
chunks(x, what="abstract")
chunks(x, what="publisher")
chunks(x, what="journal_meta")
chunks(x, what="refs_dois")
chunks(x, what=c("abstract","executive_summary"))

x <- ft_get(ids='10.1371/journal.pone.0086169', from='plos')
chunks(x, what="authors")

library("rplos")
(dois <- searchplos(q="*:*", fl='id',
   fq=list('doc_type:full',"article_type:\\"research article\\""), limit=5)$data$id)
x <- ft_get(dois, from="plos")
chunks(x, what="authors")
chunks(x, what=c("doi","categories"))
chunks(x, what="all")
chunks(x, what="publisher")
chunks(x, what="journal_meta")

# Piping workflow
opts <- list(fq=list('doc_type:full',"article_type:\\"research article\\""))
ft_search(query='ecology', from='plos', plosopts = opts)$plos$data$id \%>\%
 ft_get(from = "plos") \%>\%
 chunks("publisher")
}
}

