% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract}
\alias{extract}
\title{Extract text from a single pdf document.}
\usage{
extract(path, which, ...)
}
\arguments{
\item{path}{Path to a file}

\item{which}{One of rcamp, gs, or xpdf.}

\item{...}{further args passed on}
}
\value{
An object of class rcamp_char, gs_char, xpdf_char
}
\description{
Extract text from a single pdf document.
}
\examples{
\donttest{
res <- extract(path, "rcamp")
res
res <- extract(path, "gs")
res
res <- extract(path, "xpdf")
res
}
}

