% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_corpus}
\alias{extract_corpus}
\title{Extract text from one to many pdf documents into a tm Corpus.}
\usage{
extract_corpus(paths, which, ...)
}
\arguments{
\item{paths}{Path to a file}

\item{which}{One of rcamp, gs, or xpdf.}

\item{...}{further args passed on}
}
\value{
A tm Corpus or VCorpus
}
\description{
Extract text from one to many pdf documents into a tm Corpus.
}
\examples{
\donttest{
paths <- c("~/github/sac/scott/pdfs/BarraquandEtal2014peerj.pdf",
"~/github/sac/scott/pdfs/Chamberlain&Holland2009Ecology.pdf",
"~/github/sac/scott/pdfs/Revell&Chamberlain2014MEE.pdf")
res <- extract_corpus(paths, "rcamp")
res
tm::TermDocumentMatrix(res$data)

res <- extract_corpus(path, "gs")
res
res <- extract_corpus(path, "xpdf")
res
}
}

