% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tools.R
\name{extract_tools}
\alias{extract_tools}
\title{PDF extraction tools}
\description{
Details on the various tools for extracting pdfs in 
\pkg{fulltext}
}
\section{ft_extract}{

The \code{\link{ft_extract}} function for extracting text 
from PDFs used to allow two engines: xpdf and ghostscript. But now only 
uses Poppler via the \pkg{pdftools} package.
}

\section{ft_extract_corpus}{

The \code{\link{ft_extract_corpus}} function for extracting text 
from PDFs into a \pkg{tm} corpus used to allow two engines: xpdf and 
ghostscript. But now only uses Poppler via the \pkg{pdftools} package.
}

