% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_abstract.R
\name{ft_abstract}
\alias{ft_abstract}
\title{Get abstracts}
\usage{
ft_abstract(x, from = "scopus", scopusopts = list(), ...)
}
\arguments{
\item{x}{(character) DOIs as a character vector}

\item{from}{Source to query. only scopus for now}

\item{scopusopts}{eLife options}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
An object of class \code{ft_abstract} xxx
}
\description{
Get abstracts
}
\examples{
\dontrun{
# Scopus
opts <- list(key = Sys.getenv('ELSEVIER_SCOPUS_KEY'))
## search
(res <- ft_search(query = 'biology', from = 'scopus', scopusopts = opts, 
  limit = 25))
## get abstract
dois <- na.omit(res$scopus$data$`prism:doi`)
out <- ft_abstract(x = dois[1], scopusopts = opts)
out

(out <- ft_abstract(x = dois[1:15], scopusopts = opts))
}
}

