% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ft_get}
\alias{ft_get}
\title{Get full text.}
\usage{
ft_get(query, from = "plos", limit = 10, plosopts = list(),
  bmcopts = list(), ...)
}
\arguments{
\item{query}{Query terms}

\item{from}{Source to query}

\item{limit}{Number of records to return.}

\item{plosopts}{PLOS options. See \code{?searchplos}}

\item{bmcopts}{BMC options. See \code{?bmc_search}}

\item{...}{Further args passed on to \code{httr::GET}}
}
\value{
An object of class ft.
}
\description{
Get full text.
}
\examples{
\dontrun{
ft_search(query='ecology', from='plos')
ft_search(query='climate change', from='plos', limit=500, plosopts=list(
   fl=c('id','author','eissn','journal','counter_total_all','alm_twitterCount')))
}
}

