% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ft_get_si.R
\name{ft_get_si}
\alias{ft_get_si}
\title{Download supplementary materials from journals}
\usage{
ft_get_si(doi, si, from = c("auto", "plos", "wiley", "science", "proceedings",
  "figshare", "esa_data_archives", "esa_archives"), save.name = NULL,
  dir = NULL, cache = TRUE, vol = NULL, issue = NULL)
}
\arguments{
\item{doi}{DOI of article. Note: if using ESA journal, this must be
the ESA-specific article code (e.g., E092-201).}

\item{si}{number of the supplment to be downloads (1, 2, 3, etc.),
or (for ESA and Science journals) the name of the supplment (e.g.,
"S1_data.csv").}

\item{from}{Publisher of article. Optional, except for ESA journals
(see \code{doi}), but supplying it will speed downloads. Must be
one of: auto (i.e., auto-detect journal; default), plos, wiley,
science, proceedings, figshare, esa_data_archives, esa_archives.}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}

\item{cache}{if \code{TRUE} (default), the file won't be downloaded
again if it already exists (in a temporary directory creates, or
your chosen \code{dir})}

\item{vol}{Article volume (Proceedings journals only)}

\item{issue}{Article issue (Proceedings journals only)}
}
\description{
Download supplementary materials from papers. Put a call to this
function where you would put a file-path - everything is cached by
default, so you don't have to worry about multiple downloads in the
same session.
}
\examples{
#Put the function wherever you would put a file path
crabs <- read.csv(ft_get_si("10.6084/m9.figshare.979288", 2))

#ESA data papers and regular articles *must* be marked
fungi <- read.csv(ft_get_si("E093-059", "myco_db.csv",
                                        "esa_archives"))
mammals <- read.delim(ft_get_si("E092-201", "MCDB_communities.csv",
                                            "esa_data_archives"))
}
\author{
Will Pearse
}

