% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_init.R
\name{ft_init}
\alias{ft_init}
\title{Setup article caching}
\usage{
ft_init()
}
\value{
a \code{storr} class, invisibly
}
\description{
Setup article caching
}
\details{
the user isn't typically meant to use the output of this
function. the main use of this function is to setup a new \code{storr}
directory for caching files for any \code{\link[=ft_get]{ft_get()}} use. You can however
run this function and then interact with your storr cache yourself.
}
\examples{
\dontrun{
z <- ft_init()
class(z)
# z$destroy()
}
}
