% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ft_links.R
\name{ft_links}
\alias{ft_links}
\alias{ft_links.character}
\alias{ft_links.ft}
\title{Get full text links}
\usage{
ft_links(x, from = NULL, plosopts = list(), bmcopts = list(),
  crossrefopts = list(), entrezopts = list(), arxivopts = list(),
  biorxivopts = list(), ...)

\method{ft_links}{ft}(x, from = NULL, plosopts = list(), bmcopts = list(),
  crossrefopts = list(), entrezopts = list(), arxivopts = list(),
  biorxivopts = list(), ...)

\method{ft_links}{character}(x, from = NULL, plosopts = list(),
  bmcopts = list(), crossrefopts = list(), entrezopts = list(),
  arxivopts = list(), biorxivopts = list(), ...)
}
\arguments{
\item{x}{Query terms}

\item{from}{Source to query}

\item{plosopts}{PLOS options. See \code{?searchplos}}

\item{bmcopts}{BMC options. See \code{?bmc_search}}

\item{crossrefopts}{Crossref options. See \code{?cr_works}}

\item{entrezopts}{Entrez options. See \code{?entrez_search}}

\item{arxivopts}{arxiv options. See \code{?arxiv_search}}

\item{biorxivopts}{biorxiv options. See \code{?bx_search}}

\item{...}{Further args passed on to \code{\link[httr]{GET}}. Not working right now...}
}
\value{
An object of class ft_links
}
\description{
Get full text links
}
\examples{
\dontrun{
# Entrez
(res1 <- ft_search(query='ecology', from='entrez'))
res1$entrez$data$doi
(out <- ft_links(res1))
out$entrez
out$entrez$data[[1]]

# Crossref
(res2 <- ft_search(query='ecology', from='crossref'))
res2$crossref$data$doi
(out <- ft_links(res2))
out$crossref
out$crossref$data[[1]]

# PLOS
(res3 <- ft_search(query='ecology', from='plos', plosopts=list(
   fl=c('id','author','eissn','journal','counter_total_all','alm_twitterCount'))))
res3$plos$data$id
(out <- ft_links(res3))
out$plos
out$plos$data[[1]]

# Character input
ft_links('10.1371/journal.pone.0086169')
}
}

