% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ft_providers.R
\name{ft_providers}
\alias{ft_providers}
\title{Search for information on journals or publishers.}
\usage{
ft_providers(journal = NULL, publisher = NULL, limit = 10, ...)
}
\arguments{
\item{journal}{Query terms}

\item{publisher}{Source to query}

\item{limit}{Number of records to return.}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
An object of class ft_p
}
\description{
Search for information on journals or publishers.
}
\examples{
\dontrun{
ft_providers(journal="Stem Cells International")
ft_providers(publisher="hindawi")
ft_providers(publisher="journal")
}
}

