% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftdoi_patterns.R
\name{ftd_fetch_patterns}
\alias{ftd_fetch_patterns}
\title{Download patterns files}
\usage{
ftd_fetch_patterns()
}
\value{
character vector of file paths
}
\description{
Does various checks to see if patterns files alrady downloaded,
out of date, if some/all are deleted and in need of an update
}
\seealso{
Other ftdoi: 
\code{\link{ftd_doi}()},
\code{\link{ftd_members}()},
\code{\link{ftd_prefixes}()},
\code{\link{ftdoi_cache}},
\code{\link{prefix_local}()}
}
\concept{ftdoi}
