% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftdoi_other.R
\name{ftd_members}
\alias{ftd_members}
\title{Members}
\usage{
ftd_members(id = NULL)
}
\arguments{
\item{id}{(character) a Crossref member ID. Default is \code{NULL} which
gets all members}
}
\description{
Members
}
\examples{
\dontrun{
ftd_members()
ftd_members(221)
ftd_members(1965)

# not found
# ftd_members(999)
}
}
\seealso{
Other ftdoi: 
\code{\link{ftd_doi}()},
\code{\link{ftd_fetch_patterns}()},
\code{\link{ftd_prefixes}()},
\code{\link{ftdoi_cache}},
\code{\link{prefix_local}()}
}
\concept{ftdoi}
\keyword{internal}
