% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftxt_cache.R
\name{ftxt_cache}
\alias{ftxt_cache}
\title{Inspect and manage cached files}
\description{
Inspect and manage cached files
}
\section{Useful user functions for managing cached files}{

\itemize{
\item \code{ftxt_cache$list()} returns a character vector of full
path file names
\item \code{ftxt_cache$files()} returns file objects with metadata
\item \code{ftxt_cache$details()} returns files with details
\item \code{ftxt_cache$delete()} delete specific files
\item \code{ftxt_cache$delete_all()} delete all files, returns nothing
}
}

\examples{
\dontrun{
ftxt_cache

# list files in cache
ftxt_cache$list()

# list details of files in cache
ftxt_cache$details()

# delete certain database files
# ftxt_cache$delete("file path")
# ftxt_cache$list()

# delete all files in cache
# ftxt_cache$delete_all()
# ftxt_cache$list()
}
}
\seealso{
\link{cache}, \code{\link[=cache_file_info]{cache_file_info()}}

Other caching-functions: \code{\link{cache_file_info}},
  \code{\link{cache}}
}
\concept{caching-functions}
