% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{fulltext-deprecated}
\alias{fulltext-deprecated}
\title{Deprecated functions in fulltext}
\description{
\itemize{
\item \link{ft_chunks}: We're trying to focus the scope of the
package - this function is moving to the package \code{pubchunks}
\item \link{ft_tabularize}: We're trying to focus the scope of the
package - this function is moving to the package \code{pubchunks}
}
}
