% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltexttable.R
\docType{methods}
\name{fulltexttable}
\alias{fulltexttable}
\alias{split,fulltexttable-method}
\title{Fulltexttable class and methods}
\usage{
\S4method{split}{fulltexttable}(x, column = "tag_before", regex)
}
\arguments{
\item{x}{A \code{fulltexttable} object.}

\item{column}{The column to look up.}

\item{regex}{A regular expression.}
}
\description{
Fulltexttable class and methods
}
