% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\docType{methods}
\name{highlight,fulltexttable-method}
\alias{highlight,fulltexttable-method}
\title{Highlight tokens in fulltexttable}
\usage{
\S4method{highlight}{fulltexttable}(.Object, ...)
}
\arguments{
\item{.Object}{A \code{fulltexttable} object.}

\item{...}{See examples.}
}
\description{
Highlight tokens in fulltexttable
}
