% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_utils.R
\name{microsoft-internals}
\alias{microsoft-internals}
\alias{microsoft_search}
\alias{microsoft_links}
\title{Microsoft Academic search}
\usage{
microsoft_search(
  query,
  count = 10,
  offset = 0,
  orderby = NULL,
  atts = c("Id", "AA.AuN", "J.JN", "Ti", "Y", "E", "CC"),
  key = NULL,
  ...
)

microsoft_links(
  query,
  count = 10,
  offset = 0,
  orderby = NULL,
  atts = c("Id", "AA.AuN", "J.JN", "Ti", "Y", "E", "CC"),
  key = NULL,
  ...
)
}
\arguments{
\item{query}{(character) query terms}

\item{count}{(integer) number of records to return. default: 10}

\item{offset}{(integer) record to start at. default: 0}

\item{orderby}{(character) field to sort results by}

\item{atts}{(character) character vector of fields to return}

\item{key}{(character) microsoft academic API key, see \code{Authentication} section
in \link{fulltext-package}}
}
\description{
Wraps \code{microdemic::ma_evaluate}
}
\examples{
\dontrun{
microsoft_search(query = "Y='19'...", key = Sys.getenv("MICROSOFT_ACADEMIC_KEY"))
}
}
\references{
https://academic.microsoft.com/
}
\keyword{internal}
