% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pdfx}
\alias{pdfx}
\title{PDF-to-XML conversion of scientific articles using pdfx}
\usage{
pdfx(file = NULL, what = "parsed", ...)
}
\arguments{
\item{file}{(character) Path to a file, or files on your machine.}

\item{what}{(character) One of parsed, text, or html.}

\item{...}{Further args passed to \code{\link[httr]{GET}}. These aren't named, so just do e.g. ,
\code{verbose()}, or \code{timeout(3)}}
}
\value{
Raw XML text, parsed to XMLInternalDocument, or to html text
}
\description{
Uses a web service provided by Utopia at \url{http://pdfx.cs.man.ac.uk/}.
}
\examples{
\dontrun{
path <- "~/github/sac/hovick_work/pdfs/Vaughn1937_Mauritius.pdf"
pdfx(file = path)
}
}
\author{
Scott Chamberlain {myrmecocystus@gmail.com}
}

