% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdfx.R
\name{pdfx_targz}
\alias{pdfx_targz}
\title{Get tar.gz version of the extracted text}
\usage{
pdfx_targz(input, write_path, ...)
}
\arguments{
\item{input}{Output from \code{pdfx} function}

\item{write_path}{Path to write tar ball to.}

\item{...}{Further args passed to \code{\link[httr]{GET}}. These aren't named, so just do e.g. ,
\code{verbose()}, or \code{timeout(3)}}
}
\description{
Get tar.gz version of the extracted text
}
\examples{
\donttest{
path <- "~/github/sac/scott/pdfs/BarraquandEtal2014peerj.pdf"
out <- pdfx(file = path)
tarfile <- tempfile(fileext = "tar.gz")
pdfx_targz(input = out, write_path = tarfile)
}
}

