% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftdoi_other.R
\name{prefix_local}
\alias{prefix_local}
\title{prefix local}
\usage{
prefix_local(doi)
}
\arguments{
\item{doi}{(characte) a doi}
}
\value{
a named list with: prefix, member, name
}
\description{
prefix local
}
\examples{
\dontrun{
prefix_local('10.3390/ani4010082')
}
}
\seealso{
Other ftdoi: 
\code{\link{ftd_doi}()},
\code{\link{ftd_fetch_patterns}()},
\code{\link{ftd_members}()},
\code{\link{ftd_prefixes}()},
\code{\link{ftdoi_cache}}
}
\concept{ftdoi}
\keyword{internal}
