% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ft}
\alias{print.ft}
\title{Print brief summary of ft object}
\usage{
\method{print}{ft}(x, ..., n = 10)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}

\item{n}{Number of data frame rows to print}
}
\description{
Print brief summary of ft object
}
\examples{
\dontrun{
xxx
}
}

