% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ft_ind}
\alias{print.ft_ind}
\title{Print brief summary of ft_ind object}
\usage{
\method{print}{ft_ind}(x, ..., n = 10)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}

\item{n}{Number of data frame rows to print}
}
\description{
Print brief summary of ft_ind object
}

