% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.ft_parsed}
\alias{print.ft_parsed}
\title{Print brief summary of ft_parsed object}
\usage{
\method{print}{ft_parsed}(x, ...)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}
}
\description{
Print brief summary of ft_parsed object
}

