% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ftind}
\alias{print.ftind}
\title{Print brief summary of ftind object}
\usage{
\method{print}{ftind}(x, ..., n = 10)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}

\item{n}{Number of data frame rows to print}
}
\description{
Print brief summary of ftind object
}

