% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\docType{methods}
\name{rename}
\alias{rename}
\alias{rename,fulltexttable-method}
\alias{rename,list-method}
\title{Assign new name}
\usage{
rename(x, ...)

\S4method{rename}{fulltexttable}(x, name)

\S4method{rename}{list}(x, name)
}
\arguments{
\item{x}{Input object.}

\item{...}{Further arguments.}

\item{name}{New name to assign.}
}
\description{
Assign new name
}
