% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retag.R
\docType{methods}
\name{retag}
\alias{retag}
\alias{retag,fulltexttable-method}
\alias{retag,list-method}
\title{Retag fulltexttable}
\usage{
retag(x, ...)

\S4method{retag}{fulltexttable}(x, regex, old, new)

\S4method{retag}{list}(x, regex, old, new)
}
\arguments{
\item{x}{A \code{fulltexttable} object of a list of \code{fulltexttable} objects.}

\item{...}{Further arguments.}

\item{regex}{A regular expression to detect whether a tag should be changed.}

\item{old}{The old tag.}

\item{new}{A new tag.}
}
\description{
Retag fulltexttable
}
