% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_utils.R
\name{scopus_search}
\alias{scopus_search}
\title{Scopus search}
\usage{
scopus_search(query = NULL, type = "search", search_type = "scopus",
  key = NULL, ...)
}
\description{
Scopus search
}
\examples{
\dontrun{
scopus_search(query = "ecology")
}
}
\keyword{internal}

