% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\docType{methods}
\name{tooltips,fulltexttable-method}
\alias{tooltips,fulltexttable-method}
\title{Add tooltips to fulltexttable}
\usage{
\S4method{tooltips}{fulltexttable}(.Object, tooltips = list(), ...)
}
\arguments{
\item{.Object}{A \code{fulltexttable} object.}

\item{tooltips}{A named list or character vector with the tooltips to be
displayed. The names of the list/vector are expected to be tokens.}

\item{...}{What comes in via the three dots will be turned into a list and
merged with the tooltips argument.}
}
\description{
Add tooltips to fulltexttable
}
\examples{
\dontrun{
library(polmineR)
library(fulltext)
use("polmineR")

k <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker == "Volker Kauder") \%>\%
  subset(date == "2009-11-10") \%>\%
  as.fulltexttable(headline = "Volker Kauder (CDU)", display = "block")
  
k <- tooltips(
  k,
  Opposition = "Dauerlooser",
  Regierung = "Vollchecker",
  Regierungserklärung = "Gewinnerprogramm"
)
fulltext(k, box = FALSE)
}
}
