% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltexttable.R
\docType{methods}
\name{as.fulltexttable}
\alias{as.fulltexttable}
\alias{as.fulltexttable,slice-method}
\alias{as.fulltexttable,data.frame-method}
\alias{as.fulltexttable,character-method}
\alias{as.fulltexttable,list-method}
\alias{as.fulltexttable,fulltexttable-method}
\alias{as.fulltexttable,data.table-method}
\title{Convert object to input for fulltext (table format).}
\usage{
as.fulltexttable(x, ...)

\S4method{as.fulltexttable}{slice}(x, display = "block",
  headline = NULL, name = "", interjections = TRUE)

\S4method{as.fulltexttable}{data.frame}(x, ...)

\S4method{as.fulltexttable}{character}(x, display = "block",
  tag = "para", name = "", beautify = TRUE)

\S4method{as.fulltexttable}{list}(x, display = "block", tag = "para",
  beautify = TRUE)

\S4method{as.fulltexttable}{fulltexttable}(x, ...)

\S4method{as.fulltexttable}{data.table}(x)
}
\arguments{
\item{x}{The object to be converted.}

\item{...}{Placeholder for further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block"}

\item{headline}{A headline to prepend.}

\item{name}{An id inserted into tags.}

\item{tag}{A tag.}

\item{beautify}{Remove whitespace before interpunctation.}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
library(polmineR)
library(fulltext)
use("polmineR")
p <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
ftab <- as.fulltexttable(p, headline = "Volker Kauder (CDU)", display = "block")
fulltext(ftab, box = FALSE)

sd <- crosstalk::SharedData$new(ftab)
fulltext::fulltext(sd)
}
