% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltexttable.R
\name{as.fulltexttable}
\alias{as.fulltexttable}
\title{Convert object to input for fulltext (table format).}
\usage{
as.fulltexttable(x, headline = NULL, name = "")
}
\arguments{
\item{x}{The object to be converted.}

\item{headline}{A headline to prepend.}

\item{name}{An id inserted into tags.}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
library(polmineR)
library(fulltext)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltexttable(P, headline = "Volker Kauder (CDU)")
fulltext(D)

sd <- crosstalk::SharedData$new(D)
fulltext::fulltext(sd)
}
