% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect.ft_data}
\title{Collect data from a remote source in fulltext}
\usage{
collect(x, ...)

\method{collect}{ft_data}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Further args, ignored.}
}
\description{
Collect data from a remote source in fulltext
}
\examples{
\dontrun{
x <- ft_get(ids='10.1371/journal.pone.0086169', from='plos', cache=TRUE, backend="rds")
x \%>\% collect()
}
}

