% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/europe_pmc_utils.R
\name{eupmc}
\alias{eupmc}
\alias{eupmc_fields}
\alias{eupmc_search}
\alias{eupmc_xml}
\title{Europe PMC utilities}
\usage{
eupmc_search(query, resulttype = NULL, synonym = FALSE, per_page = 25,
  page = 1, ...)

eupmc_fields(...)

eupmc_xml(id, ...)
}
\arguments{
\item{query}{(character) Search terms. Required.}

\item{resulttype}{(character) The result type can either be idlist, core or lite.
This parameter determines the fields returned by XML and JSON formats, but it has
no effect on the DC format. See Details.}

\item{synonym}{(boolean) Synonym searches are not made by default (Default: \code{FALSE}),
however queries can be expanded using MeSH terminology and the UniProt synonyms list.
For example aspirin, a synonym would be acetylsalicylic acid;this could be included in
the search by setting the parameter value to \code{TRUE}}

\item{per_page}{(integer) Number of records to return. Max: 1000. Default: 25}

\item{page}{Page (integer) Page numbers start at 1. Default: 1}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}

\item{id}{A single Europe PMC article identifier, begins with "PMC", followed by numbers, e.g.,
"PMC3257301"}
}
\value{
\code{eupmc_search} returns a list with results. \code{eupmc_fields} returns
a data.frame. \code{eupmc_xml} returns an object of class \code{xml_document}
}
\description{
Europe PMC utilities
}
\details{
resulttype parameter options:
\itemize{
 \item idlist - returns a list of IDs and sources for the given search terms
 \item lite - returns key metadata for the given search terms; this is the default
 value if the parameter is unspecified.
 \item core - returns full metadata for a given publication ID; including abstract,
 full text links, and MeSH terms.
}
}
\examples{
\dontrun{
# search
eupmc_search(query = 'ecology')
eupmc_search(query = 'human malaria')
eupmc_search(query = '"human malaria"')
eupmc_search(query = 'auth:stoehr')
eupmc_search(query = 'journal:pnas')
eupmc_search(query = 'journal:pnas')
eupmc_search(query = 'ext_id:781840 src:med')
eupmc_search(query = 'ext_id:IND43783977 src:agr')

# list indexed search fields
head(eupmc_fields())

# get full text XML
eupmc_xml('PMC3257301')
}
}
\keyword{internal}

