% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ft_extract}
\alias{ft_extract}
\title{Extract text from a single pdf document.}
\usage{
ft_extract(path, which, ...)
}
\arguments{
\item{path}{Path to a file}

\item{which}{One of rcamp, gs, or xpdf.}

\item{...}{further args passed on}
}
\value{
An object of class rcamp_char, gs_char, xpdf_char
}
\description{
Extract text from a single pdf document.
}
\examples{
\donttest{
path <- "~/github/sac/scott/pdfs/ChamberlainEtal2013Ecosphere.pdf"

(res_rcamp <- ft_extract(path, "rcamp"))
(res_gs <- ft_extract(path, "gs"))
(res_xpdf <- ft_extract(path, "xpdf"))
}
}

