% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ft_extract_corpus.R
\name{ft_extract_corpus}
\alias{ft_extract_corpus}
\title{Extract text from one to many pdf documents into a tm Corpus or Vcorpus.}
\usage{
ft_extract_corpus(paths, which, ...)
}
\arguments{
\item{paths}{Path to one or more pdfs}

\item{which}{One of rcamp, gs, or xpdf.}

\item{...}{further args passed on}
}
\value{
A tm Corpus (or VCorpus, later that is)
}
\description{
Extract text from one to many pdf documents into a tm Corpus or Vcorpus.
}
\examples{
\donttest{
paths <- c("~/github/sac/scott/pdfs/BarraquandEtal2014peerj.pdf",
"~/github/sac/scott/pdfs/Chamberlain&Holland2009Ecology.pdf",
"~/github/sac/scott/pdfs/Revell&Chamberlain2014MEE.pdf")
(res <- ft_extract_corpus(paths, "rcamp"))
tm::TermDocumentMatrix(res$data)

(res_gs <- ft_extract_corpus(path, "gs"))
(res_xpdf <- ft_extract_corpus(path, "xpdf"))
}
}

