% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ft_get.R
\name{ft_get}
\alias{ft_get}
\title{Get full text.}
\usage{
ft_get(ids, query, from = "plos", plosopts = list(), bmcopts = list(),
  entrezopts = list(), elifeopts = list(), ...)
}
\arguments{
\item{ids}{Identifiers for papers, either DOIs, or other ids.}

\item{query}{Query terms}

\item{from}{Source to query}

\item{plosopts}{PLOS options. See \code{?searchplos}}

\item{bmcopts}{BMC options. See \code{?bmc_search}}

\item{entrezopts}{Entrez options. See \code{?entrez_search} and \code{?entrez_fetch}}

\item{elifeopts}{eLife options. See \code{?elife_search}}

\item{...}{Further args passed on to \code{httr::GET}}
}
\value{
An object of class ft_data
}
\description{
Get full text.
}
\examples{
\donttest{
ft_get(ids='10.1371/journal.pone.0086169', from='plos')
(dois <- searchplos(q="*:*", fl='id', fq='doc_type:full', limit=5)$data$id)
ft_get(ids=dois, from='plos')
ft_get(ids=c('10.7717/peerj.228','10.7717/peerj.234'), from='entrez')

# elife
ft_get(ids='10.7554/eLife.04300', from='elife')
ft_get(ids=c('10.7554/eLife.04300','10.7554/eLife.03032'), from='elife')
##### replace with code not using elife packages
dois <- searchelife(terms="Cell biology", searchin="subject_area", boolean="contains")
ft_get(ids=dois[1:10], from='elife')

# bmc
ft_get(ids='http://www.microbiomejournal.com/content/download/xml/2049-2618-2-7.xml', from='bmc')

# Frontiers in Pharmacology (publisher: Frontiers)
doi <- '10.3389/fphar.2014.00109'
ft_get(ids=doi, from="entrez")

# Hindawi Journals
ft_get(ids=c('10.1155/2014/292109','10.1155/2014/162024','10.1155/2014/249309'), from='entrez')
res <- ft_search(query='ecology', from='crossref', limit=50,
                 crossrefopts = list(filter=list(has_full_text = TRUE,
                                                 member=98,
                                                 type='journal-article')))

out <- ft_get(ids=res$crossref$data$DOI[1:20], from='entrez')
}
}

