% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ft_search}
\alias{ft_search}
\alias{print.ft}
\title{Search for full text.}
\usage{
ft_search(query, from = "plos", limit = 10, plosopts = list(),
  bmcopts = list(), ...)

\method{print}{ft}(x, ...)
}
\arguments{
\item{query}{Query terms}

\item{from}{Source to query}

\item{...}{Further args passed on to \code{httr::GET}}
}
\value{
An object of class ft.
}
\description{
Search for full text.
}
\examples{
\dontrun{
ft_search(query='ecology', from='plos')
}
}

