% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ft_search}
\alias{ft_search}
\title{Search for full text.}
\usage{
ft_search(query, from = "plos", limit = 10, plosopts = list(),
  bmcopts = list(), crossrefopts = list(), ...)
}
\arguments{
\item{query}{Query terms}

\item{from}{Source to query}

\item{limit}{Number of records to return.}

\item{plosopts}{PLOS options. See \code{?searchplos}}

\item{bmcopts}{BMC options. See \code{?bmc_search}}

\item{crossrefopts}{Crossref options. See \code{?cr_works}}

\item{...}{Further args passed on to \code{httr::GET}}
}
\value{
An object of class ft.
}
\description{
Search for full text.
}
\examples{
\dontrun{
# Plos
ft_search(query='ecology', from='plos')
ft_search(query='climate change', from='plos', limit=500, plosopts=list(
   fl=c('id','author','eissn','journal','counter_total_all','alm_twitterCount')))

# Crossref
ft_search(query='ecology', from='crossref')

# BMC
ft_search(query='ecology', from='bmc')

# Plos and Crossref
(res <- ft_search(query='ecology', from=c('plos','crossref')))
res$plos
res$crossref
}
}

