% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ft_search.R
\name{ft_search}
\alias{ft_search}
\title{Search for full text.}
\usage{
ft_search(query, from = "plos", limit = 10, plosopts = list(),
  bmcopts = list(), crossrefopts = list(), entrezopts = list(),
  arxivopts = list(), biorxivropts = list(), ...)
}
\arguments{
\item{query}{Query terms}

\item{from}{Source to query}

\item{limit}{Number of records to return.}

\item{plosopts}{PLOS options. See \code{?searchplos}}

\item{bmcopts}{BMC options. See \code{?bmc_search}}

\item{crossrefopts}{Crossref options. See \code{?cr_works}}

\item{entrezopts}{Entrez options. See \code{?entrez_search}}

\item{arxivopts}{arxiv options. See \code{?arxiv_search}}

\item{biorxivropts}{biorxivr options. See \code{?bx_search}}

\item{...}{Further args passed on to \code{\link[httr]{GET}}. Not working right now...}
}
\value{
An object of class ft, and objects of class ft_ind within each source
}
\description{
Search for full text.
}
\examples{
\dontrun{
# Plos
(res <- ft_search(query='ecology', from='plos'))
res$plos
ft_search(query='climate change', from='plos', limit=500, plosopts=list(
   fl=c('id','author','eissn','journal','counter_total_all','alm_twitterCount')))

# Crossref
(res <- ft_search(query='ecology', from='crossref'))
res$crossref

#biorxivr
(res <- ft_search(query='ecology', from='biorxivr'))
res$crossref

# BMC
(res <- ft_search(query='ecology', from='bmc'))
res$bmc

# Entrez
(res <- ft_search(query='ecology', from='entrez'))
res$entrez

# arxiv
(res <- ft_search(query='ecology', from='arxiv'))
res$arxiv

# PLOS, Crossref, and arxiv
(res <- ft_search(query='ecology', from=c('plos','crossref','arxiv')))
res$plos
res$arxiv
res$crossref
}
}

