% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ft_serialize}
\alias{ft_serialize}
\title{Serialize raw text to other formats, including to disk.}
\usage{
ft_serialize(x, to = "xml", from = NULL, ...)
}
\arguments{
\item{x}{Input object, output from a call to \code{ft_get}. Required.}

\item{to}{(character) Format to serialize to. One of xml, json, etc. Required.}

\item{from}{(character) Format \code{x} is currently in. Function attempts to use metadata
provided, or guess from data itself. Optional. CURRENTLY IGNORED.}

\item{...}{Further args passed on to \code{XML::xmlParse} or \code{jsonlite::toJSON}}
}
\description{
Serialize raw text to other formats, including to disk.
}
\examples{
\donttest{
res <- ft_get(ids='10.1371\%2Fjournal.pone.0086169', from='plos')
dois <- c('10.1371/journal.pone.0087376','10.1371\%2Fjournal.pone.0086169')
res <- ft_get(ids=dois, from='plos')
ft_serialize(res, to='json')
ft_serialize(res)
}
}

