% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{fulltext-defunct}
\alias{fulltext-defunct}
\title{Defunct functions in fulltext}
\description{
\itemize{
\item \link{ft_extract_corpus} Function removed. As part of focusing scope of the
package we're trying to limit dependencies, so downstream use of \code{tm} can
still easily be done.
\item \link{pdfx}: Function removed. We're trying to focus the scope of the
package - and this function is more out of scope now.
\item \link{chunks}: Function name changed to \code{\link[=ft_chunks]{ft_chunks()}}
\item \link{tabularize}: Function name changed to \code{\link[=ft_tabularize]{ft_tabularize()}}
\item \link{collect}: Function name changed to \code{\link[=ft_collect]{ft_collect()}}
\item \link{get_text}: Function name changed to \code{\link[=ft_text]{ft_text()}}
\item \code{cache_clear} was never working anyway, and is now removed
\item \link{ft_browse_sections}: no sign that function used, and allows
to remove a dependency
}
}
