% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\docType{methods}
\name{highlight,fulltexttable-method}
\alias{highlight,fulltexttable-method}
\title{Highlight tokens in fulltexttable}
\usage{
\S4method{highlight}{fulltexttable}(.Object, ...)
}
\arguments{
\item{.Object}{A \code{fulltexttable} object.}

\item{...}{See examples.}
}
\description{
Highlight tokens in fulltexttable
}
\examples{
library(polmineR)
library(fulltext)
use("polmineR")
kau <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker == "Volker Kauder") \%>\%
  subset(date == "2009-11-10")
f <- as.fulltexttable(kau, headline = "Volker Kauder (CDU)", display = "block")
f <- highlight(f, yellow = "Opposition")

fulltext(f, box = FALSE)
}
