% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ft_data}
\alias{print.ft_data}
\title{Print brief summary of ft_data object}
\usage{
\method{print}{ft_data}(x, ...)
}
\arguments{
\item{x}{Input...}

\item{...}{Ignored.}
}
\description{
Print brief summary of ft_data object
}
\examples{
\dontrun{
xxx
}
}

