% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_utils.R
\name{scopus_search}
\alias{scopus_search}
\title{Scopus search}
\usage{
scopus_search(query = NULL, count = 25, start = 0, type = "search",
  search_type = "scopus", key = NULL, ...)
}
\arguments{
\item{query}{query terms, as a single character vector}

\item{count}{results to return: default: 25}

\item{start}{offset value, default: 0}

\item{type}{type of search, default: search}

\item{search_type}{search type, default: scopus}

\item{key}{api key}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Scopus search
}
\examples{
\dontrun{
res <- scopus_search(query = "ecology")
res

#scopus_search(query = x, type = "abstract")

# looping through
res <- scopus_search_loop(query = "ecology community elk cow")
}
}
\keyword{internal}

